Feat: RecommendCard Alert 연결
