+++ src/pages/Profile/Profile.jsx
@@ -21,6 +21,7 @@ export default function Profile({ type }) {
   const [modalShow, setModalShow] = useState(false);
   const [modalType, setModalType] = useState("setting");
   const [selectedId, setSelectedId] = useState(null);
+  const [cardShow, setCardShow] = useState(false);
 
   function modalClose(e) {
     if (e.target === e.currentTarget) {
@@ -33,18 +34,6 @@ export default function Profile({ type }) {
     setModalType(type);
   }
 
-  const [alertShow, setAlertShow] = useState(false);
-  function alertClose(e) {
-    if (e.target === e.currentTarget) {
-      setAlertShow(false);
-    }
-  }
-
-  function alertOpen() {
-    setAlertShow(true);
-  }
-
-  const [cardShow, setCardShow] = useState(false);
   function cardClose(e) {
     if (e.target === e.currentTarget) {
       setCardShow(false);
@@ -63,9 +52,8 @@ export default function Profile({ type }) {
       <RecommendList cardOpen={cardOpen} />
       <PostList modalOpen={() => modalOpen("modification")} />
       {modalShow && (
-        <Modal type={modalType} modalClose={modalClose} alertOpen={alertOpen} />
+        <Modal type={modalType} modalClose={modalClose} modalOpen={modalOpen} />
       )}
-      {alertShow && <Alert type="logout" alertClose={alertClose} />}
       {cardShow && <RecommendCard cardClose={cardClose} id={selectedId} />}
       <Navigation />
     </Container>
